+++ Models/ModalidadeHoraExtra.cs
@@ -6,7 +6,7 @@ namespace App.Models
     {
         public int Id { get; set; }
         public string Descricao { get; set; }
-        public double Valor { get; set; }
+        public float Valor { get; set; }
         public ICollection<HoraExtra> HoraExtras { get; set; }
     }
 }
